Split structure color info out of `Common.h`
