+++ appOPHD/UI/StructureColor.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <NAS2D/Renderer/Color.h>
+
+
+enum class StructureState;
+
+
+NAS2D::Color structureColorFromIndex(StructureState structureState);
+NAS2D::Color structureTextColorFromIndex(StructureState structureState);
